refactor(runtime): split resource page reference and raw source helpers
