+++ documentation/notes/wd.codebase-overview.md
@@ -14,7 +14,7 @@ created: 1773673181726
   request/result types shared by all callers
   `core/targeting.ts` owns portable target request semantics: validate/generate target normalization, version-target normalization, prepared weave target derivation for version plus shared phases, root/recursive selection, and requested-target coverage rules including the intentional exact-versus-recursive asymmetry
   shared designator normalization now treats `/` as a CLI-only root sentinel and `""` as the internal root designator path, including root-aware target selection and support-artifact path derivation
-  `core/weave` has started splitting focused planners and shared models out of the large façade module; mesh support ResourcePage catch-up planning now lives in `mesh_support_pages.ts`, shared ResourcePage model types live in `resource_page_models.ts`, pure ResourcePage history grouping lives in `resource_page_history_groups.ts`, while `weave.ts` keeps the public re-export surface for existing runtime, CLI, and test imports
+  `core/weave` has started splitting focused planners and shared models out of the large façade module; mesh support ResourcePage catch-up planning now lives in `mesh_support_pages.ts`, shared ResourcePage model types live in `resource_page_models.ts`, pure ResourcePage history grouping lives in `resource_page_history_groups.ts`, pure ReferenceCatalog page-link extraction lives in `resource_page_reference_links.ts`, while `weave.ts` keeps the public re-export surface for existing runtime, CLI, and test imports
   first Knop, first payload, and first extracted-Knop weave planning now resolve MeshInventory current/latest/next progression from `_mesh/_meta` instead of mutable current pointers in `_mesh/_inventory`; `_mesh/_inventory` keeps stable artifact-history and historical-state membership facts while `_mesh/_meta` advances `sflo:latestHistoricalState`, `sflo:nextStateOrdinal`, and consumed next-state hints
   current carried slices: `mesh create` request validation/support-artifact rendering, `knop create` planning over an existing mesh inventory, the first narrow `integrate` planning slice for `05-alice-knop-created-woven` -> `06-alice-bio-integrated`, the first narrow `knop add-reference` planning slice for `07-alice-bio-integrated-woven` -> `08-alice-bio-referenced`, the first narrow `payload.update` planning slice for `09-alice-bio-referenced-woven` -> `10-alice-bio-updated`, `extract` planning for both Alice Bio `11-alice-bio-v2-woven` -> `12-bob-extracted` and Fantasy Rules sidecar `07-shacl-integrated-woven` -> `08-ontology-and-shacl-terms-extracted`, and carried `weave` planning slices through Alice Bio `13-bob-extracted-woven` plus Fantasy Rules sidecar `15-first-release-woven`
 
@@ -27,7 +27,7 @@ created: 1773673181726
   `runtime/config` now carries the first default effective-config seam: it loads Weave default RDF, resolves artifact-role history and ResourcePage policies, parses historical ResourcePage regeneration policy, parses default payload history/state/manifestation naming policies, parses the default config-resolution profile, and models first-pass Knop inherited-config propagation controls without changing fixture-backed behavior yet
   runtime inventory discovery, workspace loaders, and page rendering now carry the root designator path as a first-class resource when a root Knop exists at `_knop`
   runtime weave orchestration prepares a `PreparedWeaveExecution` once from shared target semantics, then dispatches version planning/writes and page generation from that prepared object instead of re-deriving generate targets from the raw request
-  `runtime/weave/weave.ts` remains the public façade for validate/version/generate/weave operations, while `runtime/weave/errors.ts`, `runtime/weave/mesh_state.ts`, `runtime/weave/planning_context.ts`, `runtime/weave/artifact_loaders.ts`, `runtime/weave/candidate_loader.ts`, `runtime/weave/request_normalization.ts`, `runtime/weave/prepared_execution.ts`, `runtime/weave/version_execution.ts`, `runtime/weave/execution_config.ts`, `runtime/weave/workspace_paths.ts`, `runtime/weave/progress.ts`, and `runtime/weave/timing_helpers.ts` now hold shared runtime errors, mesh-state loading, command-scoped staged filesystem/cache behavior, artifact-loading helpers used by both candidate loading and page context loading, candidate discovery/coverage checks, request normalization, prepared execution, version planning/writes, effective config policy adaptation, workspace path conversion, progress types, and optional timing wrappers. The planning context owns the overlay-aware read cache, dependency-aware candidate cache, planned-file staging, and cache counters used by `WEAVE_TIMING`.
+  `runtime/weave/weave.ts` remains the public façade for validate/version/generate/weave operations, while `runtime/weave/errors.ts`, `runtime/weave/mesh_state.ts`, `runtime/weave/planning_context.ts`, `runtime/weave/artifact_loaders.ts`, `runtime/weave/candidate_loader.ts`, `runtime/weave/request_normalization.ts`, `runtime/weave/prepared_execution.ts`, `runtime/weave/version_execution.ts`, `runtime/weave/execution_config.ts`, `runtime/weave/workspace_paths.ts`, `runtime/weave/progress.ts`, `runtime/weave/raw_source_panels.ts`, and `runtime/weave/timing_helpers.ts` now hold shared runtime errors, mesh-state loading, command-scoped staged filesystem/cache behavior, artifact-loading helpers used by both candidate loading and page context loading, candidate discovery/coverage checks, request normalization, prepared execution, version planning/writes, effective config policy adaptation, workspace path conversion, progress types, ResourcePage raw-source panel loading, and optional timing wrappers. The planning context owns the overlay-aware read cache, dependency-aware candidate cache, planned-file staging, and cache counters used by `WEAVE_TIMING`.
   runtime weave planning now passes Weave's default effective support-history and payload naming policies into version planning, letting first Knop and first payload weave outputs keep `_knop/_meta` current-only while preserving payload and inventory history behavior and keeping ordinal payload paths as the configured default
   runtime weave planning now passes current MeshMetadata into version planning so core can use `_mesh/_meta` as the MeshInventory progression source for the first `_mesh/_meta` migration seam
   runtime page generation and versioned inventory rendering now filter `sflo:hasResourcePage` candidates through the effective resource-page generation policy by owning artifact role, using stable history/state membership rather than mutable current/latest pointers, so `generate`, `suppress`, `defer`, and explicit-target `onRequest` have a materialization seam independent of history policy without leaving suppressed page promises in inventory
