+++ src/runtime/weave/weave.ts
@@ -3,7 +3,6 @@ import { Parser, type Quad } from "n3";
 import {
   toDesignatorResourcePagePath,
   toKnopPath,
-  toReferenceCatalogPath,
 } from "../../core/designator_segments.ts";
 import type { PlannedFile } from "../../core/planned_file.ts";
 import type {
@@ -15,7 +14,6 @@ import type {
   ResourcePageModel,
   ResourcePageRawSourcePanelModel,
   ResourcePageReferenceLinkModel,
-  ResourcePageReferenceTargetModel,
 } from "../../core/weave/resource_page_models.ts";
 import {
   collectHistoryGroupsByResourcePath
@@ -25,6 +23,7 @@ import {
   isHistoryComponentResource,
   mergeHistoryGroupsByResourcePath,
 } from "../../core/weave/resource_page_history_groups.ts";
+import { extractResourceReferenceLinks } from "../../core/weave/resource_page_reference_links.ts";
 import {
   type NormalizedTargetSpec,
   type NormalizedVersionTargetSpec,
@@ -47,8 +46,6 @@ import {
   loadOperationalLocalPathPolicy,
   LocalPathAccessError,
   type OperationalLocalPathPolicy,
-  resolveAllowedLocalPath,
-  resolveRepositorySourceFloatingLocalPath,
 } from "../operational/local_path_policy.ts";
 import { createRuntimeTiming, type RuntimeTiming } from "../timing.ts";
 import { resolveRuntimeLoggers } from "../logging/factory.ts";
@@ -57,7 +54,6 @@ import type { StructuredLogger } from "../logging/logger.ts";
 import {
   loadKnopSourceRegistryArtifact,
   loadReferenceCatalogWorkingArtifact,
-  toPayloadHistoricalSnapshotPath,
 } from "./artifact_loaders.ts";
 import { loadEffectiveConfigForExecution } from "./execution_config.ts";
 import { WeaveRuntimeError } from "./errors.ts";
@@ -79,6 +75,15 @@ import {
   normalizeVersionRequest,
   normalizeWeaveRequest,
 } from "./request_normalization.ts";
+import {
+  addCurrentKnopInventoryRawSourcePanel,
+  addExtractionSourceRawSourcePanels,
+  addPayloadRawSourcePanels,
+  addReferenceTargetSourceRawSourcePanels,
+  addSupportArtifactRawSourcePanels,
+  collectMeshSupportRawSourcePanels,
+  findRawSourcePanelsForPage,
+} from "./raw_source_panels.ts";
 import { timeOptional, timeOptionalSync } from "./timing_helpers.ts";
 import {
   prepareVersionExecution,
@@ -106,38 +111,13 @@ import {
 } from "./resource_page_policy.ts";
 import { validatePublicationPreset } from "../publication/presets.ts";
 
-const SFLO_ARTIFACT_HISTORY_IRI = `${SFLO_NAMESPACE}ArtifactHistory`;
-const SFLO_CURRENT_ARTIFACT_HISTORY_IRI =
-  `${SFLO_NAMESPACE}currentArtifactHistory`;
-const SFLO_LATEST_HISTORICAL_STATE_IRI =
-  `${SFLO_NAMESPACE}latestHistoricalState`;
-const SFLO_HAS_MANIFESTATION_IRI = `${SFLO_NAMESPACE}hasManifestation`;
-const SFLO_LOCATED_FILE_FOR_MANIFESTATION_IRI =
-  `${SFLO_NAMESPACE}locatedFileForManifestation`;
-const SFLO_LOCATED_FILE_FOR_STATE_IRI = `${SFLO_NAMESPACE}locatedFileForState`;
 const SFLO_HAS_KNOP_METADATA_IRI = `${SFLO_NAMESPACE}hasKnopMetadata`;
 const SFLO_HAS_KNOP_INVENTORY_IRI = `${SFLO_NAMESPACE}hasKnopInventory`;
 const SFLO_HAS_PAYLOAD_ARTIFACT_IRI = `${SFLO_NAMESPACE}hasPayloadArtifact`;
 const SFLO_HAS_REFERENCE_CATALOG_IRI = `${SFLO_NAMESPACE}hasReferenceCatalog`;
-const SFLO_HAS_REFERENCE_LINK_IRI = `${SFLO_NAMESPACE}hasReferenceLink`;
-const SFLO_HAS_REFERENCE_ROLE_IRI = `${SFLO_NAMESPACE}hasReferenceRole`;
-const SFLO_HAS_WORKING_LOCATED_FILE_IRI =
-  `${SFLO_NAMESPACE}hasWorkingLocatedFile`;
-const SFLO_WORKING_FILE_PATH_IRI = `${SFLO_NAMESPACE}workingLocalRelativePath`;
-const SFLO_REFERENCE_LINK_FOR_IRI = `${SFLO_NAMESPACE}referenceLinkFor`;
-const SFLO_REFERENCE_LINK_IRI = `${SFLO_NAMESPACE}ReferenceLink`;
-const SFLO_REFERENCE_ROLE_CANONICAL_IRI =
-  `${SFLO_NAMESPACE}referenceRole_canonical`;
-const SFLO_REFERENCE_TARGET_IRI = `${SFLO_NAMESPACE}referenceTarget`;
-const SFLO_REFERENCE_TARGET_STATE_IRI = `${SFLO_NAMESPACE}referenceTargetState`;
-const SFLO_REFERENCE_URI_LITERAL_IRI = `${SFLO_NAMESPACE}referenceUriLiteral`;
 const SFLO_HAS_RESOURCE_PAGE_DEFINITION_IRI =
   `${SFLO_NAMESPACE}hasResourcePageDefinition`;
 const SFLO_HAS_KNOP_ASSET_BUNDLE_IRI = `${SFLO_NAMESPACE}hasKnopAssetBundle`;
-const SFLO_ARTIFACT_RESOLUTION_MODE_WORKING_IRI =
-  `${SFLO_NAMESPACE}artifactResolutionMode_working`;
-const SFLO_ARTIFACT_RESOLUTION_MODE_LATEST_STATE_IRI =
-  `${SFLO_NAMESPACE}artifactResolutionMode_latestState`;
 const RDF_TYPE_IRI = "http://www.w3.org/1999/02/22-rdf-syntax-ns#type";
 const DCTERMS_TITLE_IRI = "http://purl.org/dc/terms/title";
 
@@ -233,15 +213,6 @@ interface GenerateDesignatorContext {
     readonly ResourcePageRawSourcePanelModel[]
   >;
 }
-
-interface ParsedResourceReferenceLink {
-  roleIris: readonly string[];
-  model: ResourcePageReferenceLinkModel;
-  referenceTargetPaths: readonly string[];
-  referenceTargetStatePaths: readonly string[];
-}
-
-const RAW_SOURCE_INLINE_BYTE_LIMIT = 1024 * 1024;
 
 export async function executeValidate(
   options: ExecuteValidateOptions,
@@ -1369,33 +1340,14 @@ async function loadGenerateDesignatorContexts(
       currentKnopInventoryTurtle,
       `Could not parse the current Knop inventory while collecting current KnopInventory source panel for ${designatorPath}.`,
     );
-    const currentKnopInventoryPagePath = `${
-      toKnopPath(designatorPath)
-    }/_inventory/index.html`;
-    const currentKnopInventoryWorkingPath = `${
-      toKnopPath(designatorPath)
-    }/_inventory/inventory.ttl`;
-    if (
-      !(await addLatestHistoricalRawSourcePanelForCurrentArtifact(
-        rawSourcePanels,
-        workspaceRoot,
-        meshState.meshBase,
-        currentKnopInventoryQuads,
-        `${toKnopPath(designatorPath)}/_inventory`,
-        currentKnopInventoryWorkingPath,
-        currentKnopInventoryPagePath,
-      ))
-    ) {
-      addRawSourcePanel(
-        rawSourcePanels,
-        currentKnopInventoryPagePath,
-        {
-          label: "Current KnopInventory file",
-          sourcePath: currentKnopInventoryWorkingPath,
-          contents: currentKnopInventoryTurtle,
-        },
-      );
-    }
+    await addCurrentKnopInventoryRawSourcePanel(
+      rawSourcePanels,
+      workspaceRoot,
+      meshState.meshBase,
+      currentKnopInventoryQuads,
+      designatorPath,
+      currentKnopInventoryTurtle,
+    );
     const resourcePageDefinitionState =
       resolveResourcePageDefinitionInventoryState(
         meshState.meshBase,
@@ -1425,6 +1377,7 @@ async function loadGenerateDesignatorContexts(
         meshState.meshBase,
         designatorPath,
         referenceCatalogArtifact.currentReferenceCatalogTurtle,
+        (message) => new WeaveRuntimeError(message),
       )
       : [];
     const sourceRegistryArtifact = await timeOptional(
@@ -1648,102 +1601,6 @@ async function loadGenerateDesignatorContexts(
   return contexts;
 }
 
-async function collectMeshSupportRawSourcePanels(
-  workspaceRoot: string,
-  meshState: MeshState,
-): Promise<ReadonlyMap<string, readonly ResourcePageRawSourcePanelModel[]>> {
-  const rawSourcePanels = new Map<
-    string,
-    readonly ResourcePageRawSourcePanelModel[]
-  >();
-  const quads = parseInventoryQuads(
-    meshState.meshBase,
-    meshState.currentMeshInventoryTurtle,
-    "Could not parse the current MeshInventory while collecting mesh support source panels.",
-  );
-
-  if (
-    !(await addLatestHistoricalRawSourcePanelForCurrentArtifact(
-      rawSourcePanels,
-      workspaceRoot,
-      meshState.meshBase,
-      quads,
-      "_mesh/_inventory",
-      "_mesh/_inventory/inventory.ttl",
-      "_mesh/_inventory/index.html",
-    ))
-  ) {
-    addRawSourcePanel(rawSourcePanels, "_mesh/_inventory/index.html", {
-      label: "Current MeshInventory file",
-      sourcePath: "_mesh/_inventory/inventory.ttl",
-      contents: meshState.currentMeshInventoryTurtle,
-    });
-  }
-
-  for (
-    const support of [
-      {
-        artifactPath: "_mesh/_meta",
-        pagePath: "_mesh/_meta/index.html",
-        sourcePath: "_mesh/_meta/meta.ttl",
-        label: "Current MeshMetadata file",
-      },
-      {
-        artifactPath: "_mesh/_config",
-        pagePath: "_mesh/_config/index.html",
-        sourcePath: "_mesh/_config/config.ttl",
-        label: "Current MeshConfig file",
-      },
-    ]
-  ) {
-    if (
-      await addLatestHistoricalRawSourcePanelForCurrentArtifact(
-        rawSourcePanels,
-        workspaceRoot,
-        meshState.meshBase,
-        quads,
-        support.artifactPath,
-        support.sourcePath,
-        support.pagePath,
-      )
-    ) {
-      continue;
-    }
-
-    try {
-      addRawSourcePanel(
-        rawSourcePanels,
-        support.pagePath,
-        await readRawSourcePanel(
-          join(workspaceRoot, support.sourcePath),
-          support.sourcePath,
-          support.label,
-        ),
-      );
-    } catch (error) {
-      if (error instanceof Deno.errors.NotFound) {
-        continue;
-      }
-      throw error;
-    }
-  }
-
-  return rawSourcePanels;
-}
-
-function findRawSourcePanelsForPage(
-  pagePath: string,
-  contexts: readonly GenerateDesignatorContext[],
-): readonly ResourcePageRawSourcePanelModel[] | undefined {
-  for (const context of contexts) {
-    const panels = context.rawSourcePanels.get(pagePath);
-    if (panels) {
-      return panels;
-    }
-  }
-  return undefined;
-}
-
 function collectKnopArtifactLinks(
   meshBase: string,
   currentKnopInventoryTurtle: string,
@@ -1878,205 +1735,6 @@ function collectHistoryGroupsByResourcePath(
   );
 }
 
-function extractResourceReferenceLinks(
-  meshBase: string,
-  designatorPath: string,
-  referenceCatalogTurtle: string,
-): readonly ParsedResourceReferenceLink[] {
-  const quads = parseInventoryQuads(
-    meshBase,
-    referenceCatalogTurtle,
-    `Could not parse the current ReferenceCatalog while collecting references for ${designatorPath}.`,
-  );
-  const referenceCatalogIri = new URL(
-    toReferenceCatalogPath(designatorPath),
-    meshBase,
-  ).href;
-  const linkSubjectPrefix = `${referenceCatalogIri}#`;
-  const designatorIri = new URL(designatorPath, meshBase).href;
-  const linkSubjects = new Set<string>();
-
-  for (const quad of quads) {
-    if (
-      quad.subject.termType !== "NamedNode" ||
-      !quad.subject.value.startsWith(linkSubjectPrefix)
-    ) {
-      continue;
-    }
-
-    const subjectIri = quad.subject.value;
-    if (
-      hasNamedNodeObject(
-        quads,
-        subjectIri,
-        RDF_TYPE_IRI,
-        SFLO_REFERENCE_LINK_IRI,
-      ) &&
-      hasNamedNodeObject(
-        quads,
-        subjectIri,
-        SFLO_REFERENCE_LINK_FOR_IRI,
-        designatorIri,
-      ) &&
-      hasNamedNodeObject(
-        quads,
-        designatorIri,
-        SFLO_HAS_REFERENCE_LINK_IRI,
-        subjectIri,
-      )
-    ) {
-      linkSubjects.add(subjectIri);
-    }
-  }
-
-  const links: ParsedResourceReferenceLink[] = [];
-  for (const subjectIri of [...linkSubjects].sort()) {
-    const roleIris = findNamedNodeObjects(
-      quads,
-      subjectIri,
-      SFLO_HAS_REFERENCE_ROLE_IRI,
-    );
-    const referenceTargetIris = findNamedNodeObjects(
-      quads,
-      subjectIri,
-      SFLO_REFERENCE_TARGET_IRI,
-    );
-    const referenceTargetStateIris = findNamedNodeObjects(
-      quads,
-      subjectIri,
-      SFLO_REFERENCE_TARGET_STATE_IRI,
-    );
-    const uriLiteralTargets = findLiteralObjects(
-      quads,
-      subjectIri,
-      SFLO_REFERENCE_URI_LITERAL_IRI,
-    );
-    const targets = toResourcePageReferenceTargets([
-      ...referenceTargetIris,
-      ...referenceTargetStateIris,
-      ...uriLiteralTargets,
-    ]);
-
-    if (roleIris.length === 0 || targets.length === 0) {
-      continue;
-    }
-
-    const referenceTargetPaths = referenceTargetIris.flatMap((iri) => {
-      const meshPath = toMeshPath(meshBase, iri);
-      return meshPath === undefined ? [] : [meshPath];
-    });
-    const referenceTargetStatePaths = referenceTargetStateIris.flatMap(
-      (iri) => {
-        const meshPath = toMeshPath(meshBase, iri);
-        return meshPath === undefined ? [] : [meshPath];
-      },
-    );
-
-    for (const roleIri of roleIris) {
-      links.push({
-        roleIris: [roleIri],
-        model: {
-          roleLabel: toReferenceRoleLabel(roleIri),
-          targets,
-        },
-        referenceTargetPaths,
-        referenceTargetStatePaths,
-      });
-    }
-  }
-
-  return links;
-}
-
-function toResourcePageReferenceTargets(
-  values: readonly string[],
-): readonly ResourcePageReferenceTargetModel[] {
-  const targets = new Map<string, ResourcePageReferenceTargetModel>();
-
-  for (const value of values) {
-    targets.set(value, {
-      href: value,
-      label: value,
-    });
-  }
-
-  return [...targets.values()].sort((left, right) =>
-    left.label.localeCompare(right.label)
-  );
-}
-
-function hasNamedNodeObject(
-  quads: readonly Quad[],
-  subjectIri: string,
-  predicateIri: string,
-  objectIri: string,
-): boolean {
-  return quads.some((quad) =>
-    quad.subject.termType === "NamedNode" &&
-    quad.subject.value === subjectIri &&
-    quad.predicate.value === predicateIri &&
-    quad.object.termType === "NamedNode" &&
-    quad.object.value === objectIri
-  );
-}
-
-function findNamedNodeObjects(
-  quads: readonly Quad[],
-  subjectIri: string,
-  predicateIri: string,
-): readonly string[] {
-  const values = new Set<string>();
-
-  for (const quad of quads) {
-    if (
-      quad.subject.termType === "NamedNode" &&
-      quad.subject.value === subjectIri &&
-      quad.predicate.value === predicateIri &&
-      quad.object.termType === "NamedNode"
-    ) {
-      values.add(quad.object.value);
-    }
-  }
-
-  return [...values].sort();
-}
-
-function findLiteralObjects(
-  quads: readonly Quad[],
-  subjectIri: string,
-  predicateIri: string,
-): readonly string[] {
-  const values = new Set<string>();
-
-  for (const quad of quads) {
-    if (
-      quad.subject.termType === "NamedNode" &&
-      quad.subject.value === subjectIri &&
-      quad.predicate.value === predicateIri &&
-      quad.object.termType === "Literal"
-    ) {
-      values.add(quad.object.value);
-    }
-  }
-
-  return [...values].sort();
-}
-
-function toReferenceRoleLabel(referenceRoleIri: string): string {
-  const localName = toLastIriSegment(referenceRoleIri);
-  const referenceRolePrefix = "referenceRole_";
-  return localName.startsWith(referenceRolePrefix)
-    ? localName.slice(referenceRolePrefix.length)
-    : localName;
-}
-
-function toLastIriSegment(iri: string): string {
-  const hashIndex = iri.lastIndexOf("#");
-  const slashIndex = iri.lastIndexOf("/");
-  const index = Math.max(hashIndex, slashIndex);
-  return index === -1 ? iri : iri.slice(index + 1);
-}
-
 function collectKnopArtifactLinksForPredicates(
   meshBase: string,
   quads: readonly Quad[],
@@ -2110,69 +1768,6 @@ function collectKnopArtifactLinksForPredicates(
   return links;
 }
 
-async function addSupportArtifactRawSourcePanels(
-  rawSourcePanels: Map<string, readonly ResourcePageRawSourcePanelModel[]>,
-  workspaceRoot: string,
-  localPathPolicy: OperationalLocalPathPolicy,
-  meshBase: string,
-  quads: readonly Quad[],
-  supportArtifacts: readonly KnopArtifactLinkModel[],
-): Promise<void> {
-  for (const artifact of supportArtifacts) {
-    const pagePath = toDesignatorResourcePagePath(artifact.path);
-    if (rawSourcePanels.has(pagePath)) {
-      continue;
-    }
-
-    const workingLocalRelativePath = resolveArtifactWorkingLocalRelativePath(
-      quads,
-      meshBase,
-      artifact.path,
-    );
-    if (!workingLocalRelativePath) {
-      continue;
-    }
-
-    if (
-      await addLatestHistoricalRawSourcePanelForCurrentArtifact(
-        rawSourcePanels,
-        workspaceRoot,
-        meshBase,
-        quads,
-        artifact.path,
-        workingLocalRelativePath,
-        pagePath,
-      )
-    ) {
-      continue;
-    }
-
-    try {
-      addRawSourcePanel(
-        rawSourcePanels,
-        pagePath,
-        await readRawSourcePanel(
-          resolveAllowedLocalPath(
-            localPathPolicy,
-            "workingLocalRelativePath",
-            workingLocalRelativePath,
-          ),
-          workingLocalRelativePath,
-          `Current ${artifact.label} file`,
-        ),
-      );
-    } catch (error) {
-      if (
-        error instanceof Deno.errors.NotFound ||
-        error instanceof LocalPathAccessError
-      ) {
-        continue;
-      }
-      throw error;
-    }
-  }
-}
-
 function findHistoryGroupsForResource(
   resourcePath: string,
   contexts: readonly GenerateDesignatorContext[],
@@ -2186,534 +1781,6 @@ function findHistoryGroupsForResource(
     }
   }
   return undefined;
-}
-
-async function addPayloadRawSourcePanels(
-  rawSourcePanels: Map<string, readonly ResourcePageRawSourcePanelModel[]>,
-  workspaceRoot: string,
-  meshBase: string,
-  quads: readonly Quad[],
-  designatorPath: string,
-  payloadArtifact: {
-    workingLocalRelativePath: string;
-    latestHistoricalStatePath?: string;
-    latestHistoricalSnapshotPath?: string;
-    latestHistoricalSnapshotTurtle?: string;
-  },
-): Promise<void> {
-  const currentPagePath = toDesignatorResourcePagePath(designatorPath);
-  await addLatestHistoricalRawSourcePanelForCurrentArtifact(
-    rawSourcePanels,
-    workspaceRoot,
-    meshBase,
-    quads,
-    designatorPath,
-    payloadArtifact.workingLocalRelativePath,
-    currentPagePath,
-    payloadArtifact.latestHistoricalSnapshotPath,
-    payloadArtifact.latestHistoricalSnapshotTurtle,
-  );
-}
-
-async function addLatestHistoricalRawSourcePanelForCurrentArtifact(
-  rawSourcePanels: Map<string, readonly ResourcePageRawSourcePanelModel[]>,
-  workspaceRoot: string,
-  meshBase: string,
-  quads: readonly Quad[],
-  artifactPath: string,
-  workingLocalRelativePath: string,
-  currentPagePath: string,
-  preloadedSnapshotPath?: string,
-  preloadedSnapshotTurtle?: string,
-): Promise<boolean> {
-  const snapshotPath = resolvePreferredLatestHistoricalLocatedFilePath(
-    quads,
-    meshBase,
-    artifactPath,
-    workingLocalRelativePath,
-  );
-  if (!snapshotPath) {
-    return false;
-  }
-
-  const latestHistoricalPanel = preloadedSnapshotPath === snapshotPath &&
-      preloadedSnapshotTurtle !== undefined
-    ? rawSourcePanelFromContents(
-      snapshotPath,
-      "Historical manifestation file",
-      preloadedSnapshotTurtle,
-    )
-    : await readHistoricalRawSourcePanel(workspaceRoot, snapshotPath);
-  if (!latestHistoricalPanel) {
-    return false;
-  }
-
-  addRawSourcePanel(
-    rawSourcePanels,
-    currentPagePath,
-    {
-      ...latestHistoricalPanel,
-      label: "Latest historical manifestation file",
-    },
-  );
-  addRawSourcePanel(
-    rawSourcePanels,
-    `${dirname(latestHistoricalPanel.sourcePath)}/index.html`,
-    latestHistoricalPanel,
-  );
-  return true;
-}
-
-async function readHistoricalRawSourcePanel(
-  workspaceRoot: string,
-  snapshotPath: string,
-): Promise<ResourcePageRawSourcePanelModel | undefined> {
-  try {
-    return await readRawSourcePanel(
-      join(workspaceRoot, snapshotPath),
-      snapshotPath,
-      "Historical manifestation file",
-    );
-  } catch (error) {
-    if (error instanceof Deno.errors.NotFound) {
-      return undefined;
-    }
-    throw error;
-  }
-}
-
-function resolvePreferredLatestHistoricalLocatedFilePath(
-  quads: readonly Quad[],
-  meshBase: string,
-  artifactPath: string,
-  workingLocalRelativePath: string,
-): string | undefined {
-  const artifactIri = new URL(artifactPath, meshBase).href;
-  const currentHistoryPath = resolveOptionalUniqueNamedNodeMeshPath(
-    quads,
-    meshBase,
-    artifactIri,
-    SFLO_CURRENT_ARTIFACT_HISTORY_IRI,
-  );
-  if (!currentHistoryPath) {
-    return undefined;
-  }
-
-  const currentHistoryIri = new URL(currentHistoryPath, meshBase).href;
-  if (
-    !hasNamedNodeObject(
-      quads,
-      currentHistoryIri,
-      RDF_TYPE_IRI,
-      SFLO_ARTIFACT_HISTORY_IRI,
-    )
-  ) {
-    return undefined;
-  }
-
-  const latestHistoricalStatePath = resolveOptionalUniqueNamedNodeMeshPath(
-    quads,
-    meshBase,
-    currentHistoryIri,
-    SFLO_LATEST_HISTORICAL_STATE_IRI,
-  );
-  if (!latestHistoricalStatePath) {
-    return undefined;
-  }
-
-  return resolvePreferredHistoricalStateLocatedFilePath(
-    quads,
-    meshBase,
-    latestHistoricalStatePath,
-    workingLocalRelativePath,
-  );
-}
-
-function resolvePreferredHistoricalStateLocatedFilePath(
-  quads: readonly Quad[],
-  meshBase: string,
-  historicalStatePath: string,
-  workingLocalRelativePath: string,
-): string | undefined {
-  const historicalStateIri = new URL(historicalStatePath, meshBase).href;
-  const locatedFilePaths = new Set<string>();
-  const shortcutLocatedFilePath = resolveOptionalUniqueNamedNodeMeshPath(
-    quads,
-    meshBase,
-    historicalStateIri,
-    SFLO_LOCATED_FILE_FOR_STATE_IRI,
-  );
-  if (shortcutLocatedFilePath) {
-    locatedFilePaths.add(shortcutLocatedFilePath);
-  }
-
-  for (
-    const manifestationIri of findNamedNodeObjects(
-      quads,
-      historicalStateIri,
-      SFLO_HAS_MANIFESTATION_IRI,
-    )
-  ) {
-    const manifestationLocatedFilePath = resolveOptionalUniqueNamedNodeMeshPath(
-      quads,
-      meshBase,
-      manifestationIri,
-      SFLO_LOCATED_FILE_FOR_MANIFESTATION_IRI,
-    );
-    if (manifestationLocatedFilePath) {
-      locatedFilePaths.add(manifestationLocatedFilePath);
-    }
-  }
-
-  return selectPreferredLocatedFilePath(
-    [...locatedFilePaths],
-    workingLocalRelativePath,
-  );
-}
-
-function selectPreferredLocatedFilePath(
-  locatedFilePaths: readonly string[],
-  workingLocalRelativePath: string,
-): string | undefined {
-  const sortedPaths = [...locatedFilePaths].sort((left, right) =>
-    left.localeCompare(right)
-  );
-  const preferredExtension = toPathExtension(workingLocalRelativePath);
-  if (preferredExtension) {
-    const extensionMatchedPath = sortedPaths.find((path) =>
-      toPathExtension(path) === preferredExtension
-    );
-    if (extensionMatchedPath) {
-      return extensionMatchedPath;
-    }
-  }
-
-  return sortedPaths[0];
-}
-
-function toPathExtension(path: string): string | undefined {
-  const fileName = toFileName(path);
-  const extensionIndex = fileName.lastIndexOf(".");
-  if (extensionIndex <= 0 || extensionIndex === fileName.length - 1) {
-    return undefined;
-  }
-
-  return fileName.slice(extensionIndex + 1).toLowerCase();
-}
-
-function rawSourcePanelFromContents(
-  sourcePath: string,
-  label: string,
-  contents: string,
-): ResourcePageRawSourcePanelModel {
-  const byteLength = new TextEncoder().encode(contents).byteLength;
-  if (byteLength > RAW_SOURCE_INLINE_BYTE_LIMIT) {
-    return {
-      label,
-      sourcePath,
-      omittedByteLength: byteLength,
-    };
-  }
-
-  return {
-    label,
-    sourcePath,
-    contents,
-  };
-}
-
-async function addReferenceTargetSourceRawSourcePanels(
-  rawSourcePanels: Map<string, readonly ResourcePageRawSourcePanelModel[]>,
-  workspaceRoot: string,
-  localPathPolicy: OperationalLocalPathPolicy,
-  meshBase: string,
-  designatorPath: string,
-  referenceLinks: readonly ParsedResourceReferenceLink[],
-): Promise<void> {
-  const seen = new Set<string>();
-
-  for (const link of referenceLinks) {
-    if (!link.roleIris.includes(SFLO_REFERENCE_ROLE_CANONICAL_IRI)) {
-      continue;
-    }
-
-    for (const referenceTargetPath of link.referenceTargetPaths) {
-      const key = `${referenceTargetPath}\u0000${
-        link.referenceTargetStatePaths.join("\u0000")
-      }`;
-      if (seen.has(key)) {
-        continue;
-      }
-      seen.add(key);
-
-      await addCanonicalReferenceSourceRawSourcePanel(
-        rawSourcePanels,
-        workspaceRoot,
-        localPathPolicy,
-        meshBase,
-        designatorPath,
-        referenceTargetPath,
-        link.referenceTargetStatePaths[0],
-      );
-    }
-  }
-}
-
-async function addCanonicalReferenceSourceRawSourcePanel(
-  rawSourcePanels: Map<string, readonly ResourcePageRawSourcePanelModel[]>,
-  workspaceRoot: string,
-  localPathPolicy: OperationalLocalPathPolicy,
-  meshBase: string,
-  designatorPath: string,
-  referenceTargetPath: string,
-  referenceTargetStatePath: string | undefined,
-): Promise<void> {
-  const sourceKnopInventoryPath = join(
-    workspaceRoot,
-    `${toKnopPath(referenceTargetPath)}/_inventory/inventory.ttl`,
-  );
-  let sourceKnopInventoryTurtle: string;
-
-  try {
-    sourceKnopInventoryTurtle = await Deno.readTextFile(
-      sourceKnopInventoryPath,
-    );
-  } catch (error) {
-    if (error instanceof Deno.errors.NotFound) {
-      return;
-    }
-    throw error;
-  }
-
-  const sourcePayloadArtifact = resolvePayloadArtifactInventoryState(
-    meshBase,
-    sourceKnopInventoryTurtle,
-    referenceTargetPath,
-    {
-      parseErrorMessage:
-        `Could not parse the canonical reference target Knop inventory while resolving page source facts for ${designatorPath}.`,
-      missingWorkingFileMessage:
-        `Could not resolve the canonical reference target working payload file for ${designatorPath}.`,
-    },
-  );
-  if (!sourcePayloadArtifact) {
-    return;
-  }
-
-  if (!referenceTargetStatePath) {
-    try {
-      addRawSourcePanel(
-        rawSourcePanels,
-        toDesignatorResourcePagePath(designatorPath),
-        await readRawSourcePanel(
-          await resolvePayloadWorkingSourcePath(
-            localPathPolicy,
-            sourcePayloadArtifact,
-          ),
-          sourcePayloadArtifact.workingLocalRelativePath,
-          "Current canonical reference source file",
-        ),
-      );
-    } catch (error) {
-      if (
-        error instanceof Deno.errors.NotFound ||
-        error instanceof LocalPathAccessError
-      ) {
-        return;
-      }
-      throw error;
-    }
-    return;
-  }
-
-  const snapshotPath = sourcePayloadArtifact.latestHistoricalStatePath ===
-        referenceTargetStatePath &&
-      sourcePayloadArtifact.latestHistoricalSnapshotPath
-    ? sourcePayloadArtifact.latestHistoricalSnapshotPath
-    : toPayloadHistoricalSnapshotPath(
-      referenceTargetStatePath,
-      sourcePayloadArtifact.workingLocalRelativePath,
-    );
-
-  try {
-    addRawSourcePanel(
-      rawSourcePanels,
-      toDesignatorResourcePagePath(designatorPath),
-      await readRawSourcePanel(
-        join(workspaceRoot, snapshotPath),
-        snapshotPath,
-        "Exact canonical reference source file",
-      ),
-    );
-  } catch (error) {
-    if (error instanceof Deno.errors.NotFound) {
-      return;
-    }
-    throw error;
-  }
-}
-
-async function addExtractionSourceRawSourcePanels(
-  rawSourcePanels: Map<string, readonly ResourcePageRawSourcePanelModel[]>,
-  workspaceRoot: string,
-  localPathPolicy: OperationalLocalPathPolicy,
-  meshBase: string,
-  designatorPath: string,
-  sourceArtifactPath: string,
-  requestedTargetStatePath: string | undefined,
-  artifactResolutionModeIri: string | undefined,
-): Promise<void> {
-  const sourceKnopInventoryPath = join(
-    workspaceRoot,
-    `${toKnopPath(sourceArtifactPath)}/_inventory/inventory.ttl`,
-  );
-  let sourceKnopInventoryTurtle: string;
-
-  try {
-    sourceKnopInventoryTurtle = await Deno.readTextFile(
-      sourceKnopInventoryPath,
-    );
-  } catch (error) {
-    if (error instanceof Deno.errors.NotFound) {
-      throw new WeaveRuntimeError(
-        `Workspace is missing the page source payload inventory for ${designatorPath}: ${
-          toKnopPath(sourceArtifactPath)
-        }/_inventory/inventory.ttl`,
-      );
-    }
-    throw error;
-  }
-
-  const sourcePayloadArtifact = resolvePayloadArtifactInventoryState(
-    meshBase,
-    sourceKnopInventoryTurtle,
-    sourceArtifactPath,
-    {
-      parseErrorMessage:
-        `Could not parse the source Knop inventory while resolving page source facts for ${designatorPath}.`,
-      missingWorkingFileMessage:
-        `Could not resolve the source working payload file for ${designatorPath}.`,
-    },
-  );
-  if (!sourcePayloadArtifact) {
-    return;
-  }
-
-  if (artifactResolutionModeIri === SFLO_ARTIFACT_RESOLUTION_MODE_WORKING_IRI) {
-    try {
-      addRawSourcePanel(
-        rawSourcePanels,
-        toDesignatorResourcePagePath(designatorPath),
-        await readRawSourcePanel(
-          await resolvePayloadWorkingSourcePath(
-            localPathPolicy,
-            sourcePayloadArtifact,
-          ),
-          sourcePayloadArtifact.workingLocalRelativePath,
-          "Working source file",
-        ),
-      );
-    } catch (error) {
-      if (
-        error instanceof Deno.errors.NotFound ||
-        error instanceof LocalPathAccessError
-      ) {
-        return;
-      }
-      throw error;
-    }
-    return;
-  }
-
-  if (
-    artifactResolutionModeIri === SFLO_ARTIFACT_RESOLUTION_MODE_LATEST_STATE_IRI
-  ) {
-    requestedTargetStatePath = sourcePayloadArtifact.latestHistoricalStatePath;
-  }
-  if (!requestedTargetStatePath) {
-    throw new WeaveRuntimeError(
-      `Extracted page source for ${designatorPath} is missing an exact target state.`,
-    );
-  }
-
-  const snapshotPath = sourcePayloadArtifact.latestHistoricalStatePath ===
-        requestedTargetStatePath &&
-      sourcePayloadArtifact.latestHistoricalSnapshotPath
-    ? sourcePayloadArtifact.latestHistoricalSnapshotPath
-    : toPayloadHistoricalSnapshotPath(
-      requestedTargetStatePath,
-      sourcePayloadArtifact.workingLocalRelativePath,
-    );
-
-  try {
-    addRawSourcePanel(
-      rawSourcePanels,
-      toDesignatorResourcePagePath(designatorPath),
-      await readRawSourcePanel(
-        join(workspaceRoot, snapshotPath),
-        snapshotPath,
-        "Exact source file",
-      ),
-    );
-  } catch (error) {
-    if (error instanceof Deno.errors.NotFound) {
-      throw new WeaveRuntimeError(
-        `Workspace is missing the exact page source file for ${designatorPath}: ${snapshotPath}`,
-      );
-    }
-    throw error;
-  }
-}
-
-async function resolvePayloadWorkingSourcePath(
-  localPathPolicy: OperationalLocalPathPolicy,
-  payloadArtifact: {
-    workingLocalRelativePath: string;
-    repositorySourceFloatingLocator?: RepositorySourceFloatingLocator;
-  },
-): Promise<string> {
-  return payloadArtifact.repositorySourceFloatingLocator
-    ? await resolveRepositorySourceFloatingLocalPath(
-      localPathPolicy,
-      payloadArtifact.repositorySourceFloatingLocator,
-    )
-    : resolveAllowedLocalPath(
-      localPathPolicy,
-      "workingLocalRelativePath",
-      payloadArtifact.workingLocalRelativePath,
-    );
-}
-
-async function readRawSourcePanel(
-  absolutePath: string,
-  sourcePath: string,
-  label: string,
-): Promise<ResourcePageRawSourcePanelModel> {
-  const info = await Deno.stat(absolutePath);
-  if (info.size > RAW_SOURCE_INLINE_BYTE_LIMIT) {
-    return {
-      label,
-      sourcePath,
-      omittedByteLength: info.size,
-    };
-  }
-
-  return {
-    label,
-    sourcePath,
-    contents: await Deno.readTextFile(absolutePath),
-  };
-}
-
-function addRawSourcePanel(
-  rawSourcePanels: Map<string, readonly ResourcePageRawSourcePanelModel[]>,
-  pagePath: string,
-  panel: ResourcePageRawSourcePanelModel,
-): void {
-  rawSourcePanels.set(pagePath, [
-    ...(rawSourcePanels.get(pagePath) ?? []),
-    panel,
-  ]);
 }
 
 function describeSemanticFlowResource(
@@ -2888,83 +1955,6 @@ function toLastPathSegment(path: string): string {
   return segments[segments.length - 1] ?? "/";
 }
 
-function resolveArtifactWorkingLocalRelativePath(
-  quads: readonly Quad[],
-  meshBase: string,
-  artifactPath: string,
-): string | undefined {
-  const artifactIri = new URL(artifactPath, meshBase).href;
-  const literalValue = resolveOptionalUniqueLiteralObject(
-    quads,
-    artifactIri,
-    SFLO_WORKING_FILE_PATH_IRI,
-  );
-  const locatedFileValue = resolveOptionalUniqueNamedNodeMeshPath(
-    quads,
-    meshBase,
-    artifactIri,
-    SFLO_HAS_WORKING_LOCATED_FILE_IRI,
-  );
-
-  if (
-    literalValue !== undefined &&
-    locatedFileValue !== undefined &&
-    literalValue !== locatedFileValue
-  ) {
-    return undefined;
-  }
-
-  return literalValue ?? locatedFileValue;
-}
-
-function resolveOptionalUniqueLiteralObject(
-  quads: readonly Quad[],
-  subjectIri: string,
-  predicateIri: string,
-): string | undefined {
-  const values = new Set<string>();
-  for (const quad of quads) {
-    if (
-      quad.subject.termType !== "NamedNode" ||
-      quad.subject.value !== subjectIri ||
-      quad.predicate.value !== predicateIri ||
-      quad.object.termType !== "Literal"
-    ) {
-      continue;
-    }
-
-    values.add(quad.object.value);
-  }
-
-  return values.size === 1 ? values.values().next().value! : undefined;
-}
-
-function resolveOptionalUniqueNamedNodeMeshPath(
-  quads: readonly Quad[],
-  meshBase: string,
-  subjectIri: string,
-  predicateIri: string,
-): string | undefined {
-  const values = new Set<string>();
-  for (const quad of quads) {
-    if (
-      quad.subject.termType !== "NamedNode" ||
-      quad.subject.value !== subjectIri ||
-      quad.predicate.value !== predicateIri ||
-      quad.object.termType !== "NamedNode"
-    ) {
-      continue;
-    }
-
-    const meshPath = toMeshPath(meshBase, quad.object.value);
-    if (meshPath !== undefined) {
-      values.add(meshPath);
-    }
-  }
-
-  return values.size === 1 ? values.values().next().value! : undefined;
-}
-
 function toMeshPath(meshBase: string, iri: string): string | undefined {
   const meshUrl = new URL(meshBase);
   const iriUrl = new URL(iri);
@@ -2999,11 +1989,6 @@ function toResourcePath(pagePath: string): string {
   return pagePath.endsWith("/index.html")
     ? pagePath.slice(0, -"/index.html".length)
     : pagePath;
-}
-
-function toFileName(path: string): string {
-  const segments = path.split("/");
-  return segments[segments.length - 1]!;
 }
 
 const GENERATED_TIMESTAMP_FOOTER_PATTERN =
